+++ platform/src/codatplatform/models/operations/update_connection_authorization.py
@@ -4,6 +4,7 @@ from __future__ import annotations
 import dataclasses
 import requests as requests_http
 from ..shared import connection as shared_connection
+from ..shared import errormessage as shared_errormessage
 from typing import Dict, Optional
 
 
@@ -26,6 +27,8 @@ class UpdateConnectionAuthorizationResponse:
     r"""HTTP response status code for this operation"""
     connection: Optional[shared_connection.Connection] = dataclasses.field(default=None)
     r"""OK"""
+    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
+    r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""