+++ platform/RELEASES.md
@@ -39,3 +39,13 @@ Based on:
 - [python v2.0.0] platform
 ### Releases
 - [PyPI v2.0.0] https://pypi.org/project/codat-platform/2.0.0 - platform
+
+## 2023-11-06 01:28:40
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.112.1 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.1.0] platform
+### Releases
+- [PyPI v2.1.0] https://pypi.org/project/codat-platform/2.1.0 - platform